Lab2-Assignment1
Time to declare a large array on heap structure is much longer than on
stack. Because the heap is a more complicated data structrue. Allocating
memory on stack is just a single instruction about changing the pointer
of stack. But on heap, it requires big enough block and  need to insure
that which part of memory is available.